fix exception msg

+++ clients/src/main/java/org/apache/kafka/clients/producer/KafkaProducer.java
@@ -1061,11 +1061,11 @@ public class KafkaProducer<K, V> implements Producer<K, V> {
         float ratio = CompressionRatioEstimator.estimation(topic, compressionType);
         float expectedSize = size * ratio;
         if (expectedSize > maxRequestSize)
-            throw new RecordTooLargeException("The message expected size using compression type " + compressionType + " is " + expectedSize +
+            throw new RecordTooLargeException("The message expected size using compression type " + compressionType + " is " + (int) expectedSize +
                 " bytes when serialized which is larger than " + maxRequestSize + ", which is the value of the " +
                 ProducerConfig.MAX_REQUEST_SIZE_CONFIG + " configuration. (original size = " + size + " and expected ratio = " + ratio + ")");
         if (expectedSize > totalMemorySize)
-            throw new RecordTooLargeException("The message expected size using compression type " + compressionType + " is " + expectedSize +
+            throw new RecordTooLargeException("The message expected size using compression type " + compressionType + " is " + (int) expectedSize +
                 " bytes when serialized which is larger than the total memory buffer you have configured with the " +
                 ProducerConfig.BUFFER_MEMORY_CONFIG +
                 " configuration. (original size = " + size + " and expected ratio = " + ratio + ")");
